requirements.txt
# Django Core
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-environ==0.11.2

# Database
psycopg2-binary==2.9.10
dj-database-url==2.2.0

# Authentication & Security
djangorestframework-simplejwt==5.4.0
django-allauth==65.3.0

# API & Forms
django-crispy-forms==2.3
crispy-bootstrap5==2025.6

# Task Queue (para automações futuras)
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0

# Integrations & Payments
stripe==11.3.0
requests==2.32.3

# AI & Data Analysis
openai==1.59.7
anthropic==0.42.0
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0

# Utilities
python-decouple==3.8
whitenoise==6.8.2
gunicorn==23.0.0
python-dateutil==2.9.0.post0

# Development
django-debug-toolbar==4.4.6
django-extensions==3.2.3
